+++ e3nn_jax/_src/irreps_array.py
@@ -608,24 +608,24 @@ class IrrepsArray:
         keep: Union[e3nn.Irreps, List[e3nn.Irrep], Callable[[e3nn.MulIrrep], bool]] = None,
         *,
         drop: Union[e3nn.Irreps, List[e3nn.Irrep], Callable[[e3nn.MulIrrep], bool]] = None,
+        lmax: int = None,
     ) -> "IrrepsArray":
         r"""Filter the irreps.
 
         Args:
             keep (Irreps or list of `Irrep` or function): list of irrep to keep
             exclude (Irreps or list of `Irrep` or function): list of irrep to exclude
+            lmax (int): maximum l
 
         Examples:
-            >>> IrrepsArray("0e + 2x1o + 2x0e", jnp.arange(9)).filtered(["1o"])
+            >>> IrrepsArray("0e + 2x1o + 2x0e", jnp.arange(9)).filter(["1o"])
             2x1o [1 2 3 4 5 6]
         """
-        if keep is None and drop is None:
+        if keep is None and drop is None and lmax is None:
             return self
-        if keep is not None and drop is not None:
-            raise ValueError("Cannot specify both keep and drop")
 
         backend = _infer_backend(self.array)
-        new_irreps = self.irreps.filter(keep=keep, drop=drop)
+        new_irreps = self.irreps.filter(keep=keep, drop=drop, lmax=lmax)
         return IrrepsArray.from_list(
             new_irreps,
             [x for x, mul_ir in zip(self.list, self.irreps) if mul_ir in new_irreps],